Dived cairo-coverage to crates
commit-id:2f1053a5
