+++ crates/cairo-coverage-args/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "cairo-coverage-args"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+camino.workspace = true
+clap.workspace = true
