+++ crates/cairo-coverage-core/src/coverage_data/function.rs
