+++ crates/cairo-coverage-core/src/data_loader/loaded_data.rs
