+++ crates/cairo-coverage-core/src/input/data.rs
