+++ crates/cairo-coverage-core/src/input/sierra_to_cairo_map.rs
