+++ crates/cairo-coverage-core/src/merge.rs
