+++ crates/cairo-coverage-core/src/output/lcov.rs
