+++ crates/cairo-coverage-core/src/types.rs
