+++ crates/cairo-coverage/Cargo.toml
@@ -4,22 +4,14 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+cairo-coverage-args = { path = "../cairo-coverage-args" }
+cairo-coverage-core = { path = "../cairo-coverage-core" }
 anyhow.workspace = true
-camino.workspace = true
-cairo-annotations.workspace = true
-cairo-lang-sierra.workspace = true
-cairo-lang-sierra-to-casm.workspace = true
-cairo-lang-starknet-classes.workspace = true
 clap.workspace = true
-derived-deref.workspace = true
-itertools.workspace = true
-ignore.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-regex.workspace = true
-indoc.workspace = true
+walkdir.workspace = true
+
 
 [dev-dependencies]
+cairo-coverage-test-utils = { path = "../cairo-coverage-test-utils" }
 assert_fs.workspace = true
 snapbox.workspace = true
-
